Add files via upload
Task#1:
Suppose that billing Amount is a variable, which denotes the amount you need to pay to
the department store. If you pay the full amount, you get $10.00 or 1% of the billing
Amount, whichever is smaller, as a credit on your next bill; if you pay at least 50% of the
billing Amount, the penalty is 5% of the balance; if you pay at least 20% of the billing
Amount and less than 50% of the billing Amount, the penalty is 10% of the balance;
otherwise, the penalty is 20% of the balance. Design an algorithm that prompts the user
to enter the billing amount and the desired payment. The algorithm then calculates and
outputs the credit or the remaining balance. If the amount is not paid in full, the algorithm
should also output the penalty amount.

Task#2:
Write a program to remove the integer 20 from the list given below using loops.

Task#3:
You have given a list of task 2. Write a program to replace the integer 20 with 200. Only
replace the first occurrence.

Task#4:
Write a python program to check if all items in the tuple are same

Task#5:
Write a program that prints the numbers from 1 to 40 (10 numbers in one line, total of 4
lines). But multiples of 4 should be replaced with your first name, multiples of 5 should be
replaced with your second name, and multiples of both 4 and 5 should be replaced with
your roll number? USING 1 FOR LOOP